refactor: add package entrypoint
