+++ README.md
@@ -18,7 +18,9 @@ docker compose up --build
 
 ```bash
 pip install -e .
-python -m web_app.server
+python -m docrouter
+# или через созданный скрипт
+docrouter
 ```
 
 После старта интерфейс будет доступен по адресу [http://localhost:8000](http://localhost:8000), где можно загружать, просматривать и скачивать документы.
