+++ pyproject.toml
@@ -26,6 +26,9 @@ dependencies = [
     "python-magic",
 ]
 
+[project.scripts]
+docrouter = "docrouter.main:main"
+
 [tool.setuptools]
 package-dir = {"" = "src"}
 py-modules = [
