+++ src/docrouter/__main__.py
@@ -1,0 +1,5 @@
+"""CLI запуск для DocRouter."""
+from .main import main
+
+if __name__ == "__main__":
+    main()
